Help command, usage parameters in commands

+++ commands/Fun/say.js
@@ -1,11 +1,12 @@
 module.exports = {
     name: 'say',
-    description: 'Makes it look like as if the bot said something',
-    event: 'messageCreate',
+    description: 'Makes it look like as if the bot said something.',
+    usage: '`b!say <message>`',
     execute(message){
         console.log('Going into say');
         
         const botSay = message.content.slice(6);
+        if(!botSay) return message.reply(`C'mon you need to give me something to say, burunyu!`)
         const channel = message.channel;
 
         message.delete();
